Perform a GET request to Tremendous API

This function provides lower-level access to perform GET requests via Tremendous API. Available endpoints can be found on the official Tremendous API documentation.

trem_get(
  client,
  path,
  query = list(),
  disk = NULL,
  stream = NULL,
  api_key,
  sandbox,
  parse = TRUE,
  ...
)

Arguments

client
A Tremendous API Client object, created with trem_client_new.

path
The URL path, appended to the base URL, for GET requests such as listing available payment types, funding sources, account members, and more. See the Tremendous API Documentation for examples.

query
Query terms as a named list. See crul::HttpClient for more details.

disk
A path to write to. NULL by default, so info is written to memory. See crul::HttpClient for more details.

stream
An R function to determine how to stream data. NULL by default, so info is streaned with memory. See crul::HttpClient for more details.

api_key
API key from tremendous.com. Can either pass in here as a character string or set for repeated use with trem_set_api_key.

sandbox
Logical: TRUE (default) and any API requests are performed within the Tremendous sandbox environment, a free and fully-featured environment for application developing and testing. FALSE and the API requests are performed within the Tremendous production environment. This will involve sending actual money, so be certain you wish to do this!

parse
Logical: Should the API Response results be parsed into a data frame?

...
Curl options passed to crul::verb-GET

Value

If parse = TRUE (default), a list containing the response from the API request. Otherwise, the R6 HttpResponse object containing API request data.

Examples