R6 Class representing a new Tremendous API Client

R6 Class representing a new Tremendous API Client

R6 Class representing a new Tremendous API Client

Details

Called by trem_client_new to bundle API key and environment.

Public fields

key

(character) an API key

sandbox

(logical) tremClient for Sandbox environment (TRUE) or production (FALSE)?

httpClient

for internal use

Methods

Public methods

tremendous_client$new()

tremendous_client$print()

tremendous_client$clone()

Method new()

Create a new tremClient object

Usage

tremendous_client$new(
  api_key = NULL,
  sandbox = TRUE,
  curl_opts = list(useragent = trem_ua())
)

Arguments

api_key

API key from tremendous.com. Can either pass in here as a character string or set for repeated use with trem_set_api_key.

sandbox

Logical: TRUE (default) and any API requests are performed within the Tremendous sandbox environment, a free and fully-featured environment for application developing and testing. FALSE and the API requests are performed within the Tremendous production environment. This will involve sending actual money, so be certain you wish to do this!

curl_opts

A named list of curl options for the API Client. Defaults to include useragent info.`

Method print()

Printing method for object of class 'tremClient'.

Usage

tremendous_client$print(...)

Arguments

...

NA; printing function

Method clone()

The objects of this class are cloneable with this method.

Usage

tremendous_client$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.